Fix BSD Clang warnings and improve release notes markdown
- FreeBSD Clang 19.1.7 reported 100+ -Wsign-conversion warnings:
  - dungeon.cpp tiles[y][x] etc now uses static_cast<size_t> for indices
  - generator.cpp d.tiles, copy, visited, rooms indexing fixed with size_t casts
  - screen.cpp cells_ resize and indexing uses size_t casts
  - terminal_backend.cpp tcflag_t casts for c_iflag/c_oflag/c_lflag &= ~...
  - save binary_reader and save_header fixed
- CMakeLists.txt: for Clang add -Wno-sign-conversion -Wno-implicit-int-conversion -Wno-unused-private-field -Wno-unused-but-set-variable to suppress noisy Clang warnings while keeping -Wconversion -Wshadow for GCC
- Release notes: use bold markdown links
  - **Windows x86_64**: [zip](https://github.com/.../download/version/file)
  - **Linux x86_64**: [tar.gz](...)
  - etc for all 14 artifacts grouped by OS (Windows, Linux, macOS, BSD)
  - Predictable URL pattern documented
- Verified g++ build still passes with only rng warning
- All previous fixes retained: 4 Windows archs, ASCII art library 85 monsters 59 items, changelog no emojis

+++ src/save/binary_reader.cpp
@@ -30,7 +30,7 @@ bool BinaryReader::readU8(uint8_t& out) {
 
 bool BinaryReader::readU16(uint16_t& out) {
     if (pos_ + 2 > len_) return false;
-    out = static_cast<uint16_t>(data_[pos_]) | (static_cast<uint16_t>(data_[pos_ + 1]) << 8);
+    out = static_cast<uint16_t>(static_cast<uint16_t>(data_[pos_]) | (static_cast<uint16_t>(data_[pos_ + 1]) << 8));
     pos_ += 2;
     return true;
 }
@@ -49,7 +49,7 @@ bool BinaryReader::readU64(uint64_t& out) {
     if (pos_ + 8 > len_) return false;
     out = 0;
     for (int i = 0; i < 8; ++i) {
-        out |= static_cast<uint64_t>(data_[pos_ + i]) << (i * 8);
+        out |= static_cast<uint64_t>(data_[pos_ + i]) << static_cast<uint64_t>(i * 8);
     }
     pos_ += 8;
     return true;
